Changed Scanner to BufferedReader; added descriptions for 4 first tasks; restructured project;
