+++ src/tasks/Third.java
@@ -1,0 +1,4 @@
+package tasks;
+
+public class Third {
+}
